+++ test/test_primitives.py
@@ -23,13 +23,14 @@ from qiskit.primitives import (
     BaseSamplerV1,
     Sampler,
 )
-from qiskit.providers import Backend, BackendV2
+from qiskit.providers import Backend
 from qiskit.quantum_info import SparsePauliOp
 from qiskit.transpiler.exceptions import TranspilerError
 
+from qiskit_aqt_provider.aqt_resource import AnyAQTResource
 from qiskit_aqt_provider.primitives import AQTSampler
 from qiskit_aqt_provider.primitives.estimator import AQTEstimator
-from qiskit_aqt_provider.test.circuits import assert_circuits_equal
+from qiskit_aqt_provider.test.circuits import assert_circuits_equal, random_circuit
 from qiskit_aqt_provider.test.fixtures import MockSimulator
 
 
@@ -66,7 +67,8 @@ def test_backend_primitives_are_v1() -> None:
 )
 @pytest.mark.httpx_mock(assert_all_responses_were_requested=False)
 def test_circuit_sampling_primitive(
-    get_sampler: Callable[[Backend], BaseSamplerV1], any_offline_simulator_no_noise: BackendV2
+    get_sampler: Callable[[Backend], BaseSamplerV1],
+    any_offline_simulator_no_noise: AnyAQTResource,
 ) -> None:
     """Check that a `Sampler` primitive using an AQT backend can sample parametric circuits."""
     theta = Parameter("θ")
@@ -183,3 +185,22 @@ def test_aqt_sampler_transpilation(theta: float, offline_simulator_no_noise: Moc
     tr_expected = qiskit.transpile(expected, offline_simulator_no_noise)
 
     assert_circuits_equal(transpiled_circuit, tr_expected)
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)  # for the direct-access mocks
+def test_sampler_circuit_batching(any_offline_simulator_no_noise: AnyAQTResource) -> None:
+    """Check that a Sampler primitive on an offline simulator can split oversized job batches.
+
+    Regression test for #203.
+    """
+    # Arbitrary circuit.
+    qc = random_circuit(2)
+
+    sampler = AQTSampler(any_offline_simulator_no_noise)
+
+    # Use a Sampler batch larger than the maximum number of circuits
+    # per batch in the API.
+    batch_size = 3 * any_offline_simulator_no_noise.max_circuits
+    result = sampler.run([qc] * batch_size).result()
+
+    assert len(result.quasi_dists) == batch_size